feat(log): adds logging support
adds basic logging using the crate 'log' so that
downstream apps can see more information about what
is going on under the hood.

primarily just using 'info!' for anytime settingsfile
attempts to access the filesystem, this helps with most
of the errors and issues (accessing files, dirs that doen't
exist).